Pin AWS provider to the major version we are currently using
Version 3 of the AWS provider was recently rolled out, and we want to
avoid it since it is a hot mess right now.

Projects that inherit from this skeleton should pin any other
providers they use to the major version currently in use.  This will
help us avoid unwelcome surprises in the future.