Update the terraform-docs configuration
Add a key to support outputting ATX closed headers which is our
preference. This is also required to make updating the README(s) more
convenient.

NOTE:
This functionality is entirely dependent on the code in the
`improvement/support_atx_closed_markdown_headers` branch of
mcdonnnj/terraform-docs. There is a PR (745) to the original repository
but it has not been merged. Please note that I'm not using auto-linking
language to avoid inundating the PR with links due to this commit being
re-used in all downstream repositories.